# server/openapi.yaml
openapi: 3.1.0
info:
  title: Fake StackOverflow REST API
  version: 1.0.0
  description: |
    API for managing questions, answers, tags, and users of a fake stack overflow application.

    The API is organized into four main routes:
    * Tags route: Managing tags and their statistics
    * Questions route: Managing questions and their operations
    * Answers route: Managing answers to questions
    * Users route: Managing user registration, login, and user profiles
servers:
  - url: http://localhost:8000
    description: Development server for the Fake StackOverflow REST API.
...
components:
  responses:
    BadRequest:
      description: Bad Request - Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Internal Server Error"
    Forbidden:
      description: Forbidden - User is not authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Forbidden"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not authenticated"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                example: ".body.name"
              message:
                type: string
                example: "must be string"
              errorCode:
                type: string
                example: "type.openapi.validation"
...
paths:
...
  /comment/add:
      post:
        tags:
          - Comments
        summary: Add a new comment to a post
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - postId
                  - postType
                  - text
                properties:
                  postId:
                    type: string
                    example: "6641e6b0d75a1a6ac8b5cc11"
                  postType:
                    type: string
                    enum: [Question, Answer]
                    example: "Question"
                  text:
                    type: string
                    example: "This is a comment."
        responses:
          "200":
            description: Comment added successfully
            content:
              application/json:
                schema:
                  type: object
          "400":
            $ref: "#/components/responses/BadRequest"
          "500":
            $ref: "#/components/responses/InternalServerError"
          "403":
            $ref: "#/components/responses/Forbidden"
          "401":
            $ref: "#/components/responses/Unauthorized"
  /comment/edit:
    post:
      tags:
        - Comments
      summary: Edit an existing comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - _id
                - postId
                - postType
                - text
              properties:
                _id:
                  type: string
                  example: "6641e6b0d75a1a6ac8b5cc99"
                postId:
                  type: string
                  example: "6641e6b0d75a1a6ac8b5cc11"
                postType:
                  type: string
                  enum: [Question, Answer]
                  example: "Question"
                text:
                  type: string
                  example: "Updated comment text"
      responses:
        "200":
          description: Comment edited successfully
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "403":
            $ref: "#/components/responses/Forbidden"
        "401":
            $ref: "#/components/responses/Unauthorized"
  /comment/delete/{id}:
    post:
      tags:
        - Comments
      summary: Soft delete a comment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "6641e6b0d75a1a6ac8b5cc99"
      responses:
        "200":
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "403":
            $ref: "#/components/responses/Forbidden"
        "401":
            $ref: "#/components/responses/Unauthorized"
  /comment/{postType}/{postId}:
    get:
      tags:
        - Comments
      summary: Get all comments for a post (question or answer)
      parameters:
        - name: postType
          in: path
          required: true
          schema:
            type: string
            enum: [Question, Answer]
            example: "Question"
        - name: postId
          in: path
          required: true
          schema:
            type: string
            example: "6641e6b0d75a1a6ac8b5cc11"
      responses:
        "200":
          description: Successfully retrieved comments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"